requirements.txt
wandb
pytorch_lightning==2.1.3
torch_geometric==2.5.0
librosa==0.10.1
einops==0.7.0
audiomentations==0.35.0
openai-whisper
tqdm==4.66.2
numpy==1.26.4